src/requirements.txt
paho-mqtt==1.4.0
pytz
datetime
requests
